requirements.txt
# Backend Framework
Flask==3.0.0
flask-cors==4.0.0
flasgger==0.9.7.1

# Base de Datos
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

# OCR (FUNCIONAN EN WINDOWS)
pytesseract==0.3.10
easyocr==1.7.1

# Procesamiento PDFs
PyMuPDF==1.23.8
pdfplumber==0.10.3
pypdf==3.17.1

# Procesamiento Imágenes
opencv-python==4.9.0.80
Pillow==9.5.0
scikit-image==0.22.0

# Excel
openpyxl==3.1.2
xlrd==2.0.1

# IA (compatible en Windows)
transformers==4.36.0

# Análisis de Datos
pandas==2.2.2
polars==0.20.2
numpy==1.26.4

# Utilidades
chardet==5.2.0
python-dotenv==1.0.0

# Servidor
gunicorn==21.2.0
waitress==2.1.2
